// src/redux/action-types.tsx
export enum ViewActions {
    SET_PAGE = 'SET_PAGE',
    SET_EDIT_MODE = 'SET_EDIT_MODE',
    SET_DRAWER_VISIBLE = 'SET_DRAWER_VISIBLE',
    SET_DIALOG_VISIBLE = 'SET_DIALOG_VISIBLE',
}

export enum PreferenceActions {
    SET_LEFT_RIGHT_PADDING = 'SET_LEFT_RIGHT_PADDING',
    SET_TOP_BOTTOM_PADDING = 'SET_TOP_BOTTOM_PADDING',
    SET_SPELLCHECK = 'SET_SPELL_CHECK',
}

export enum WordCardActions {
    SET_WORD = 'SET_WORD',
    SET_CARD_VISIBLE = 'SET_CARD_VISIBLE',
    SET_MEANING_EXAMPLES = 'SET_MEANING_EXAMPLES',
}